requirements.txt

Django~=4.2.14
bcrypt
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.2.0